@@ -15,51 +15,50 @@
     "postsetup": "cd src && (rename savi-lib lib || mv savi-lib lib) && cd lib && (git remote add upstream https://github.com/ml5174/savi-lib.git || echo skipping, moving on) && git checkout master && git pull upstream master && cd ../.. && (git remote add upstream https://github.com/ml5174/sacpcp.git || echo skipping, moving on) && git checkout master && git pull upstream master && npm install -g ionic cordova && npm install && ionic serve"
   },
   "dependencies": {
-    "@angular/animations": "4.4.3",
-    "@angular/common": "4.4.3",
-    "@angular/compiler": "4.4.3",
-    "@angular/compiler-cli": "4.4.3",
-    "@angular/core": "4.4.3",
-    "@angular/forms": "4.4.3",
-    "@angular/http": "4.4.3",
-    "@angular/platform-browser": "4.4.3",
-    "@angular/platform-browser-dynamic": "4.4.3",
-    "@ionic-native/app-version": "^4.4.2",
-    "@ionic-native/core": "4.3.0",
-    "@ionic-native/dialogs": "^4.4.2",
-    "@ionic-native/in-app-browser": "4.3.1",
-    "@ionic-native/splash-screen": "4.3.0",
-    "@ionic-native/status-bar": "4.3.0",
-    "@ionic/storage": "2.0.1",
-    "@ngx-translate/core": "6.0.1",
-    "@ngx-translate/http-loader": "^2.0.0",
-    "angular2-text-mask": "^8.0.4",
+    "@angular/animations": "5.2.7",
+    "@angular/common": "5.2.7",
+    "@angular/compiler": "5.2.7",
+    "@angular/compiler-cli": "5.2.7",
+    "@angular/core": "5.2.7",
+    "@angular/forms": "5.2.7",
+    "@angular/http": "5.2.7",
+    "@angular/platform-browser": "5.2.7",
+    "@angular/platform-browser-dynamic": "5.2.7",
+    "@ionic-native/app-version": "^4.5.3",
+    "@ionic-native/core": "4.5.3",
+    "@ionic-native/dialogs": "4.5.3",
+    "@ionic-native/in-app-browser": "4.5.3",
+    "@ionic-native/splash-screen": "^4.5.3",
+    "@ionic-native/status-bar": "^4.5.3",
+    "@ionic/storage": "2.1.3",
+    "@ngx-translate/core": "9.1.1",
+    "@ngx-translate/http-loader": "^2.0.1",
+    "angular2-text-mask": "8.0.4",
     "cordova-android": "6.3.0",
     "cordova-ios": "4.5.2",
-    "cordova-plugin-app-version": "^0.1.9",
+    "cordova-plugin-appversion": "^1.0.0",
     "cordova-plugin-device": "^1.1.7",
-    "cordova-plugin-dialogs": "^1.3.4",
-    "cordova-plugin-inappbrowser": "^1.7.2",
-    "cordova-plugin-ionic-webview": "^1.1.15",
+    "cordova-plugin-dialogs": "^2.0.0",
+    "cordova-plugin-inappbrowser": "^2.0.2",
+    "cordova-plugin-ionic-webview": "^1.1.16",
     "cordova-plugin-splashscreen": "^4.1.0",
-    "cordova-plugin-statusbar": "git+https://github.com/apache/cordova-plugin-statusbar.git",
+    "cordova-plugin-statusbar": "2.4.1",
     "cordova-plugin-whitelist": "^1.3.3",
-    "ionic-angular": "3.7.1",
-    "ionic-native": "^2.9.0",
+    "ionic-angular": "^3.9.2",
     "ionic-plugin-deeplinks": "~1.0.15",
     "ionic-plugin-keyboard": "^2.2.1",
     "ionicons": "3.0.0",
-    "moment": "^2.19.3",
-    "moment-timezone": "^0.5.10",
-    "rxjs": "5.4.3",
+    "moment": "2.20.1",
+    "moment-timezone": "0.5.14",
+    "rxjs": "^5.5.6",
     "sw-toolbox": "3.6.0",
     "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@angular/tsc-wrapped": "^4.4.4",
-    "@ionic/app-scripts": "^3.1.4",
-    "@types/moment-timezone": "^0.2.32",
-    "typescript": "2.3.4"
+    "@angular/tsc-wrapped": "^4.4.6",
+    "@ionic/app-scripts": "^3.1.8",
+    "@types/moment-timezone": "^0.5.4",
+    "typescript": "~2.4.2"
   },
   "description": "savi: An Ionic project",
   "cordova": {
