@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
-import 'rxjs/add/operator/map';
+//import { map } from 'rxjs/add/operator/map';
 import { SERVER } from '../lib/provider/config';
 
 
